- コンボボックスを選択した際の文字色を独立して設定する機能を追加 - メーターの文字列が溢れたときに浮動小数点表示を諦めて整数表示する機能を追加 - ベンチマークを途中でやめた際にレイテンシ測定結果が 0 となる不具合を修正

+++ ButtonCx.cpp
@@ -343,7 +343,7 @@ void CButtonCx::DrawString(CDC *drawDC, LPDRAWITEMSTRUCT lpDrawItemStruct)
 			}
 			else
 			{
-				x = rect.left;
+				x = (REAL)rect.left;
 			}
 
 			if (x < 0)
